refactor: static_cast in BalancePD, named PWM constants, remove stray imu.cpp

+++ src/pid/pid_control.cpp
@@ -10,7 +10,7 @@ int BalancePD::compute(float angle, float gyro)
 {
     float angle_bias = mid_angle_ - angle;
     float gyro_bias = 0 - gyro;
-    return -kp_ / 100.0f * angle_bias - gyro_bias * kd_ / 100.0f;
+    return static_cast<int>(-kp_ / 100.0f * angle_bias - gyro_bias * kd_ / 100.0f);
 }
 
 // Velocity PI
